% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/disaggregatedFinancialStatements_200_Liabilities.R
\docType{data}
\name{disaggregatedFinancialStatements_200_Liabilities}
\alias{disaggregatedFinancialStatements_200_Liabilities}
\title{Disaggregated Financial Statements Liabilities 200}
\format{
Data frame with columns
\describe{
\item{CD_200_00}{Liabilities}
\item{CD_200_XX}{Sub Liabilities}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_200_Liabilities
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_200_Liabilities
}
\keyword{datasets}
