% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregatedFinancialStatements_300_Equity.R
\docType{data}
\name{disaggregatedFinancialStatements_300_Equity}
\alias{disaggregatedFinancialStatements_300_Equity}
\title{Disaggregated Financial Statements Equity 300}
\format{
Data frame with columns
\describe{
\item{CD_300_00}{Equity}
\item{CD_300_XX}{Sub Equity}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_300_Equity
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_300_Equity
}
\keyword{datasets}
