% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregatedFinancialStatements_400_Expenses.R
\docType{data}
\name{disaggregatedFinancialStatements_400_Expenses}
\alias{disaggregatedFinancialStatements_400_Expenses}
\title{Disaggregated Financial Statements Expenses 400}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_400_Expenses
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_400_Expenses
}
\keyword{datasets}
