% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregatedFinancialStatements_500_Income.R
\docType{data}
\name{disaggregatedFinancialStatements_500_Income}
\alias{disaggregatedFinancialStatements_500_Income}
\title{Disaggregated Financial Statements Income 500}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_500_Income
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_500_Income
}
\keyword{datasets}
