% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/disaggregatedFinancialStatements_700_ContingentCreditor.R
\docType{data}
\name{disaggregatedFinancialStatements_700_ContingentCreditor}
\alias{disaggregatedFinancialStatements_700_ContingentCreditor}
\title{Disaggregated Financial Statements Contingent Creditor 700}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_700_ContingentCreditor
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_700_ContingentCreditor
}
\keyword{datasets}
