% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/disaggregatedFinancialStatements_900_DebtorCreditor.R
\docType{data}
\name{disaggregatedFinancialStatements_900_DebtorCreditor}
\alias{disaggregatedFinancialStatements_900_DebtorCreditor}
\title{Disaggregated Financial Statements Debtor Creditor 900}
\format{
Data frame with columns
\describe{
\item{CD_900_00}{Debtor Creditor}
\item{CD_900_XX}{Sub Debtor Creditor}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_900_DebtorCreditor
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_900_DebtorCreditor
}
\keyword{datasets}
