% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregatedFinancialStatementsAccess.R
\docType{data}
\name{disaggregatedFinancialStatementsAccess}
\alias{disaggregatedFinancialStatementsAccess}
\title{Disaggregated Financial Statements Access}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatementsAccess
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatementsAccess
}
\keyword{datasets}
