% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/disaggregatedFinancialStatementsAccountsManual.R
\docType{data}
\name{disaggregatedFinancialStatementsAccountsManual}
\alias{disaggregatedFinancialStatementsAccountsManual}
\title{disaggregatedFinancialStatementsAccountsManual}
\format{
Data frame with columns
\describe{
\item{CODIGO}{Acess}
\item{DESCRIPCION}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatementsAccountsManual
}
\description{
The table records the asset accounts
}
\examples{
  disaggregatedFinancialStatementsAccountsManual
}
\keyword{datasets}
