% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicatorsByFinancialInstitutions.R
\docType{data}
\name{indicatorsByFinancialInstitutions}
\alias{indicatorsByFinancialInstitutions}
\title{indicatorsByFinancialInstitutions}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
indicatorsByFinancialInstitutions
}
\description{
The table records the asset accounts
}
\examples{
  indicatorsByFinancialInstitutions
}
\keyword{datasets}
