% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/loanPortfolioPlacementsByTypeOfEconomicActivityDebtor.R
\docType{data}
\name{loanPortfolioPlacementsByTypeOfEconomicActivityDebtor}
\alias{loanPortfolioPlacementsByTypeOfEconomicActivityDebtor}
\title{loanPortfolioPlacementsByTypeOfEconomicActivityDebtor}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
loanPortfolioPlacementsByTypeOfEconomicActivityDebtor
}
\description{
The table records the asset accounts
}
\examples{
  loanPortfolioPlacementsByTypeOfEconomicActivityDebtor
}
\keyword{datasets}
