% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loanPortfolioPlacementsByTypeOfLoan.R
\docType{data}
\name{loanPortfolioPlacementsByTypeOfLoan}
\alias{loanPortfolioPlacementsByTypeOfLoan}
\title{loanPortfolioPlacementsByTypeOfLoan}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
loanPortfolioPlacementsByTypeOfLoan
}
\description{
The table records the asset accounts
}
\examples{
  loanPortfolioPlacementsByTypeOfLoan
}
\keyword{datasets}
