% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/loanPortfolioPlacementsByTypeOfLoanDestination.R
\docType{data}
\name{loanPortfolioPlacementsByTypeOfLoanDestination}
\alias{loanPortfolioPlacementsByTypeOfLoanDestination}
\title{loanPortfolioPlacementsByTypeOfLoanDestination}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
loanPortfolioPlacementsByTypeOfLoanDestination
}
\description{
The table records the asset accounts
}
\examples{
  loanPortfolioPlacementsByTypeOfLoanDestination
}
\keyword{datasets}
