% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFinancialStatements.R
\docType{data}
\name{summaryFinancialStatements}
\alias{summaryFinancialStatements}
\title{Summary Financial Statements}
\format{
Data frame with columns
\describe{
\item{XXXXX}{Account}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
summaryFinancialStatements
}
\description{
The table of financial statements corresponding
to the summarized financial statements of the
different financial institutions in Bolivia.
}
\examples{
  summaryFinancialStatements
}
\keyword{datasets}
