% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticketsFinancialInstitutions.R
\docType{data}
\name{ticketsFinancialInstitutions}
\alias{ticketsFinancialInstitutions}
\title{ticketsFinancialInstitutions}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
ticketsFinancialInstitutions
}
\description{
The table records the asset accounts
}
\examples{
  ticketsFinancialInstitutions
}
\keyword{datasets}
