% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregatedFinancialStatements_100_Assets.R
\docType{data}
\name{disaggregatedFinancialStatements_100_Assets}
\alias{disaggregatedFinancialStatements_100_Assets}
\title{Disaggregated Financial Statements Assets 100}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_100_Assets
}
\description{
The table records the asset accounts of the
disaggregated financial statements of the
institutions of the Bolivian financial system.
}
\examples{
  disaggregatedFinancialStatements_100_Assets
}
\keyword{datasets}
