% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/disaggregatedFinancialStatements_600_ContingentDebtor.R
\docType{data}
\name{disaggregatedFinancialStatements_600_ContingentDebtor}
\alias{disaggregatedFinancialStatements_600_ContingentDebtor}
\title{Disaggregated Financial Statements Contingent Debtor 600}
\format{
Data frame with columns
\describe{
\item{CD_600_00}{Contingent Debtor}
\item{CD_600_XX}{Sub Contingent Debtor}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_600_ContingentDebtor
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_600_ContingentDebtor
}
\keyword{datasets}
