% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/disaggregatedFinancialStatements_800_DebtorOrder.R
\docType{data}
\name{disaggregatedFinancialStatements_800_DebtorOrder}
\alias{disaggregatedFinancialStatements_800_DebtorOrder}
\title{Disaggregated Financial Statements Debtor Order 800}
\format{
Data frame with columns
\describe{
\item{CD_800_00}{Debtor Order}
\item{CD_800_XX}{Sub Debtor Order}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
disaggregatedFinancialStatements_800_DebtorOrder
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  disaggregatedFinancialStatements_800_DebtorOrder
}
\keyword{datasets}
