% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFinancialStatements.R
\docType{data}
\name{summaryFinancialStatements}
\alias{summaryFinancialStatements}
\title{summaryFinancialStatements}
\format{
Data frame with columns
\describe{
\item{CD_100_00}{Acess}
\item{CD_100_XX}{Sub Access}
}
}
\source{
\url{https://www.asfi.gob.bo/}
}
\usage{
summaryFinancialStatements
}
\description{
The table records the asset accounts
}
\examples{
  summaryFinancialStatements
}
\keyword{datasets}
